requirements.txt
python-telegram-bot[webhooks]==20.8
Flask==2.3.3
